+++ data/src/main/java/com/suwiki/data/datasource/remote/RemoteRestrictionHistoryDataSource.kt
@@ -1,0 +1,11 @@
+package com.suwiki.data.datasource.remote
+
+import com.suwiki.model.Result
+import com.suwiki.model.Suspension
+
+interface RemoteRestrictionHistoryDataSource {
+
+    suspend fun getSuspensionHistory(): Result<List<Suspension.Ban>>
+
+    suspend fun getBlacklistHistory(): Result<List<Suspension.Block>>
+}
