+++ remote/src/main/java/com/suwiki/remote/api/AuthApi.kt
@@ -1,6 +1,7 @@
 package com.suwiki.remote.api
 
 import com.suwiki.remote.ApiResult
+import com.suwiki.remote.api.UserApi.Companion.USER
 import com.suwiki.remote.request.user.CheckEmailRequest
 import com.suwiki.remote.request.user.CheckIdRequest
 import com.suwiki.remote.request.user.FindIdRequest
@@ -23,33 +24,8 @@ import retrofit2.http.POST
 // TODO : v2 api로 업그레이드 필요.
 interface AuthApi {
     companion object {
-        const val USER = "/user"
         const val AUTH_HEADER = "Authorization"
     }
-
-    // 회원가입 요청 API
-    @POST("$USER/join")
-    suspend fun signUp(@Body signupRequest: SignupRequest): ApiResult<SuccessCheckResponse>
-
-    // 아이디 중복 확인 요청 API
-    @POST("$USER/check-id")
-    suspend fun checkId(@Body checkIdRequest: CheckIdRequest): ApiResult<OverlapCheckResponse>
-
-    // 이메일 중복 확인 요청 API
-    @POST("$USER/check-email")
-    suspend fun checkEmail(@Body checkEmailRequest: CheckEmailRequest): ApiResult<OverlapCheckResponse>
-
-    // 아이디 찾기 API
-    @POST("$USER/find-id")
-    suspend fun findId(@Body findIdRequest: FindIdRequest): ApiResult<SuccessCheckResponse>
-
-    // 비밀번호 찾기(임시 비밀번호 전송) API
-    @POST("$USER/find-pw")
-    suspend fun findPassword(@Body findPasswordRequest: FindPasswordRequest): ApiResult<SuccessCheckResponse>
-
-    // 로그인 요청 API
-    @POST("$USER/login")
-    suspend fun login(@Body loginRequest: LoginRequest): ApiResult<TokenResponse>
 
     // Refresh Token
     @POST("$USER/refresh")
