+++ remote/src/main/java/com/suwiki/remote/datasource/RemoteUserDataSourceImpl.kt
@@ -2,55 +2,28 @@ package com.suwiki.remote.datasource
 
 import com.suwiki.data.datasource.remote.RemoteUserDataSource
 import com.suwiki.model.Result
-import com.suwiki.model.Suspension
 import com.suwiki.model.Token
 import com.suwiki.model.User
-import com.suwiki.remote.api.AuthApi
 import com.suwiki.remote.api.UserApi
-import com.suwiki.remote.request.user.CheckEmailRequest
-import com.suwiki.remote.request.user.CheckIdRequest
 import com.suwiki.remote.request.user.FindIdRequest
 import com.suwiki.remote.request.user.FindPasswordRequest
 import com.suwiki.remote.request.user.LoginRequest
 import com.suwiki.remote.request.user.QuitRequest
 import com.suwiki.remote.request.user.ResetPasswordRequest
-import com.suwiki.remote.request.user.SignupRequest
 import com.suwiki.remote.response.user.toModel
 import com.suwiki.remote.toResult
 import javax.inject.Inject
 
 class RemoteUserDataSourceImpl @Inject constructor(
     private val userApi: UserApi,
-    private val authApi: AuthApi,
 ) : RemoteUserDataSource {
-    override suspend fun signUp(id: String, password: String, email: String): Result<Boolean> {
-        val request = SignupRequest(
-            id = id,
-            password = password,
-            email = email,
-        )
-
-        return authApi.signUp(
-            signupRequest = request,
-        ).toResult().map { it.success }
-    }
-
-    override suspend fun checkId(loginId: String): Result<Boolean> {
-        return authApi.checkId(CheckIdRequest(loginId)).toResult().map {
-            it.overlap
-        }
-    }
-
-    override suspend fun checkEmail(email: String): Result<Boolean> {
-        return authApi.checkEmail(CheckEmailRequest(email)).toResult().map { it.overlap }
-    }
 
     override suspend fun findId(email: String): Result<Boolean> {
-        return authApi.findId(FindIdRequest(email)).toResult().map { it.success }
+        return userApi.findId(FindIdRequest(email)).toResult().map { it.success }
     }
 
     override suspend fun findPassword(loginId: String, email: String): Result<Boolean> {
-        return authApi.findPassword(
+        return userApi.findPassword(
             FindPasswordRequest(loginId, email),
         ).toResult().map { it.success }
     }
@@ -68,16 +41,12 @@ class RemoteUserDataSourceImpl @Inject constructor(
     }
 
     override suspend fun login(loginId: String, password: String): Result<Token> {
-        return authApi.login(
+        return userApi.login(
             LoginRequest(
                 loginId = loginId,
                 password = password,
             ),
         ).toResult().map { it.toModel() }
-    }
-
-    override fun requestRefresh(refresh: String): Result<Token> {
-        return authApi.requestRefresh(refresh).toResult().map { it.toModel() }
     }
 
     override suspend fun quit(id: String, password: String): Result<Boolean> {
@@ -91,15 +60,5 @@ class RemoteUserDataSourceImpl @Inject constructor(
 
     override suspend fun getUserData(): Result<User> {
         return userApi.getUserData().toResult().map { it.toModel() }
-    }
-
-    override suspend fun getSuspensionHistory(): Result<List<Suspension.Ban>> {
-        return userApi.getSuspensionHistory().toResult()
-            .map { result -> result.map { it.toModel() } }
-    }
-
-    override suspend fun getBlacklistHistory(): Result<List<Suspension.Block>> {
-        return userApi.getBlacklistHistory().toResult()
-            .map { result -> result.map { it.toModel() } }
     }
 }
